Test: Refactor and adjust property tests for Expr/Prop
We move property tests for Expr/Prop to a dedicated ExprTests module.
We also move generators for these types to a separate module.
