+++ test/test.hs
@@ -8,7 +8,6 @@ module Main where
 import Prelude hiding (LT, GT)
 
 import GHC.TypeLits
-import Data.Proxy
 import Control.Monad
 import Control.Monad.ST (stToIO)
 import Control.Monad.State.Strict
@@ -21,7 +20,6 @@ import Data.ByteString.Base16 qualified as BS16
 import Data.ByteString.Lazy qualified as BSLazy
 import Data.Binary.Put (runPut)
 import Data.Binary.Get (runGetOrFail)
-import Data.DoubleWord
 import Data.Either
 import Data.List qualified as List
 import Data.Map.Strict qualified as Map
@@ -34,9 +32,8 @@ import Data.Text qualified as T
 import Data.Time (diffUTCTime, getCurrentTime)
 import Data.Tuple.Extra
 import Data.Tree (flatten)
-import Data.Typeable
 import Data.Vector qualified as V
-import Data.Word (Word8, Word64)
+import Data.Word (Word8)
 import Data.Char (digitToInt)
 import GHC.Conc (getNumProcessors)
 import System.Directory
@@ -63,7 +60,7 @@ import EVM.Assembler
 import EVM.Exec
 import EVM.Expr qualified as Expr
 import EVM.Fetch qualified as Fetch
-import EVM.Format (hexByteString, hexText, formatExpr)
+import EVM.Format (hexByteString, hexText)
 import EVM.Precompiled
 import EVM.RLP
 import EVM.SMT hiding (one)
@@ -540,213 +537,6 @@ tests = testGroup "hevm"
         -- there won't be query now as accessStorage uses fetch cache
         assertBoolM (show vm4.result) (isNothing vm4.result)
     ]
-  -- These tests fuzz the simplifier by generating a random expression,
-  -- applying some simplification rules, and then using the smt encoding to
-  -- check that the simplified version is semantically equivalent to the
-  -- unsimplified one
-  , adjustOption (\(Test.Tasty.QuickCheck.QuickCheckTests n) -> Test.Tasty.QuickCheck.QuickCheckTests (div n 2)) $ testGroup "SimplifierPropertyTests"
-    [ test "read-beyond-bound (negative-test)" $ do
-      let
-        e1 = CopySlice (Lit 1) (Lit 0) (Lit 2) (ConcreteBuf "a") (ConcreteBuf "")
-        e2 = ConcreteBuf "Definitely not the same!"
-      equal <- checkEquiv e1 e2
-      assertBoolM "Should not be equivalent!" $ not equal
-    , testProperty  "buffer-simplification" $ \(expr :: Expr Buf) -> propNoSimp $ do
-        let simplified = Expr.simplify expr
-        checkEquivAndLHS expr simplified
-    , testProperty  "buffer-simplification-len" $ \(expr :: Expr Buf) -> propNoSimp $ do
-        let simplified = Expr.simplify (BufLength expr)
-        checkEquivAndLHS (BufLength expr) simplified
-    , testProperty "store-simplification" $ \(expr :: Expr Storage) -> propNoSimp $ do
-        let simplified = Expr.simplify expr
-        checkEquivAndLHS expr simplified
-    , testProperty "load-simplification" $ \(GenWriteStorageLoad expr) -> propNoSimp $ do
-        let simplified = Expr.simplify expr
-        checkEquivAndLHS expr simplified
-    , ignoreTest $ testProperty "load-decompose" $ \(GenWriteStorageLoad expr) -> propNoSimp $ do
-        putStrLnM $ T.unpack $ formatExpr expr
-        let simp = Expr.simplify expr
-        let decomposed = fromMaybe simp $ mapExprM Expr.decomposeStorage simp
-        -- putStrLnM $ "-----------------------------------------"
-        -- putStrLnM $ T.unpack $ formatExpr decomposed
-        -- putStrLnM $ "\n\n\n\n"
-        checkEquiv expr decomposed
-    , testProperty "byte-simplification" $ \(expr :: Expr Byte) -> propNoSimp $ do
-        let simplified = Expr.simplify expr
-        checkEquivAndLHS expr simplified
-    , askOption $ \(QuickCheckTests n) -> testProperty "word-simplification" $ withMaxSuccess (min n 20) $ \(ZeroDepthWord expr) -> propNoSimp $ do
-        let simplified = Expr.simplify expr
-        checkEquivAndLHS expr simplified
-    , testProperty "readStorage-equivalance" $ \(store, slot) -> propNoSimp $ do
-        let simplified = Expr.readStorage' slot store
-            full = SLoad slot store
-        checkEquiv full simplified
-    , testProperty "writeStorage-equivalance" $ \(val, GenWriteStorageExpr (slot, store)) -> propNoSimp $ do
-        let simplified = Expr.writeStorage slot val store
-            full = SStore slot val store
-        checkEquiv full simplified
-    , testProperty "readWord-equivalance" $ \(buf, idx) -> propNoSimp $ do
-        let simplified = Expr.readWord idx buf
-            full = ReadWord idx buf
-        checkEquiv full simplified
-    , testProperty "writeWord-equivalance" $ \(idx, val, WriteWordBuf buf) -> propNoSimp $ do
-        let simplified = Expr.writeWord idx val buf
-            full = WriteWord idx val buf
-        checkEquiv full simplified
-    , testProperty "arith-simplification" $ \(_ :: Int) -> propNoSimp $ do
-        expr <- liftIO $ generate . sized $ genWordArith 15
-        let simplified = Expr.simplify expr
-        checkEquivAndLHS expr simplified
-    , testProperty "readByte-equivalance" $ \(buf, idx) -> propNoSimp $ do
-        let simplified = Expr.readByte idx buf
-            full = ReadByte idx buf
-        checkEquiv full simplified
-    -- we currently only simplify concrete writes over concrete buffers so that's what we test here
-    , testProperty "writeByte-equivalance" $ \(LitOnly val, LitOnly buf, GenWriteByteIdx idx) -> propNoSimp $ do
-        let simplified = Expr.writeByte idx val buf
-            full = WriteByte idx val buf
-        checkEquiv full simplified
-    , testProperty "copySlice-equivalance" $ \(srcOff, GenCopySliceBuf src, GenCopySliceBuf dst, LitWord @300 size) -> propNoSimp $ do
-        -- we bias buffers to be concrete more often than not
-        dstOff <- liftIO $ generate (maybeBoundedLit 100_000)
-        let simplified = Expr.copySlice srcOff dstOff size src dst
-            full = CopySlice srcOff dstOff size src dst
-        checkEquiv full simplified
-    , testProperty "indexWord-equivalence" $ \(src, LitWord @50 idx) -> propNoSimp $ do
-        let simplified = Expr.indexWord idx src
-            full = IndexWord idx src
-        checkEquiv full simplified
-    , testProperty "pow-base2-simp" $ \(_ :: Int) -> propNoSimp $ do
-        expo <- liftIO $ generate . sized $ genWordArith 15
-        let full = Exp (Lit 2) expo
-            simplified = Expr.simplify full
-        checkEquiv full simplified
-    , testProperty "pow-low-exponent-simp" $ \(LitWord @100 expo) -> propNoSimp $ do
-        base <- liftIO $ generate . sized $ genWordArith 15
-        let full = Exp base expo
-            simplified = Expr.simplify full
-        checkEquiv full simplified
-    , testProperty "indexWord-mask-equivalence" $ \(src :: Expr EWord, LitWord @35 idx) -> propNoSimp $ do
-        mask <- liftIO $ generate $ do
-          pow <- arbitrary :: Gen Int
-          frequency
-           [ (1, pure $ Lit $ (shiftL 1 (pow `mod` 256)) - 1)        -- potentially non byte aligned
-           , (1, pure $ Lit $ (shiftL 1 ((pow * 8) `mod` 256)) - 1)  -- byte aligned
-           ]
-        let
-          input = And mask src
-          simplified = Expr.indexWord idx input
-          full = IndexWord idx input
-        checkEquiv full simplified
-    , testProperty "toList-equivalance" $ \buf -> propNoSimp $ do
-        let
-          -- transforms the input buffer to give it a known length
-          fixLength :: Expr Buf -> Gen (Expr Buf)
-          fixLength = mapExprM go
-            where
-              go :: Expr a -> Gen (Expr a)
-              go = \case
-                WriteWord _ val b -> liftM3 WriteWord idx (pure val) (pure b)
-                WriteByte _ val b -> liftM3 WriteByte idx (pure val) (pure b)
-                CopySlice so _ sz src dst -> liftM5 CopySlice (pure so) idx (pure sz) (pure src) (pure dst)
-                AbstractBuf _ -> cbuf
-                e -> pure e
-              cbuf = do
-                bs <- arbitrary
-                pure $ ConcreteBuf bs
-              idx = do
-                w <- arbitrary
-                -- we use 100_000 as an upper bound for indices to keep tests reasonably fast here
-                pure $ Lit (w `mod` 100_000)
-
-        input <- liftIO $ generate $ fixLength buf
-        case Expr.toList input of
-          Nothing -> do
-            putStrLnM "skip"
-            pure True -- ignore cases where the buf cannot be represented as a list
-          Just asList -> do
-            let asBuf = Expr.fromList asList
-            checkEquiv asBuf input
-    , testProperty "simplifyProp-equivalence-lit" $ \(LitProp p) -> propNoSimp $ do
-        let simplified = Expr.simplifyProps [p]
-        case simplified of
-          [] -> checkEquivProp (PBool True) p
-          [val@(PBool _)] -> checkEquivProp val p
-          _ -> liftIO $ assertFailure "must evaluate down to a literal bool"
-    , testProperty "simplifyProp-equivalence-sym" $ \(p) -> propNoSimp $ do
-        let simplified = Expr.simplifyProp p
-        checkEquivPropAndLHS p simplified
-    , testProperty "simplify-joinbytes" $ \(SymbolicJoinBytes exprList) -> propNoSimp $ do
-        let x = joinBytesFromList exprList
-        let simplified = Expr.simplify x
-        y <- checkEquiv x simplified
-        assertBoolM "Must be equal" y
-    , testProperty "simpProp-equivalence-sym-Prop" $ \(ps :: [Prop]) -> propNoSimp $ do
-        let simplified = pand (Expr.simplifyProps ps)
-        checkEquivPropAndLHS (pand ps) simplified
-    , testProperty "simpProp-equivalence-sym-LitProp" $ \(LitProp p) -> propNoSimp $ do
-        let simplified = pand (Expr.simplifyProps [p])
-        checkEquivPropAndLHS p simplified
-    , testProperty "storage-slot-simp-property" $ \(StorageExp s) -> propNoSimp $ do
-        -- we have to run `Expr.litToKeccak` on the unsimplified system, or
-        -- we'd need some form of minimal simplifier for things to work out. As long as
-        -- we trust the litToKeccak, this is fine, as that function is stand-alone,
-        -- and quite minimal
-        let s2 = Expr.litToKeccak s
-        let simplified = Expr.simplify s2
-        checkEquivAndLHS s2 simplified
-    , test "storage-slot-single" $ do
-        -- this tests that "" and "0"x32 is not equivalent in Keccak
-        let x = SLoad (Add (Keccak (ConcreteBuf "")) (Lit 1)) (SStore (Keccak (ConcreteBuf "\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL")) (Lit 0) (AbstractStore (SymAddr "stuff") Nothing))
-        let simplified = Expr.simplify x
-        y <- checkEquivAndLHS x simplified
-        assertBoolM "Must be equal" y
-    , test "word-eq-bug" $ do
-        -- This test is actually OK because the simplified takes into account that it's impossible to find a
-        -- near-collision in the keccak hash
-        let x =  (SLoad (Keccak (AbstractBuf "es")) (SStore (Add (Keccak (ConcreteBuf "")) (Lit 0x1)) (Lit 0xacab) (ConcreteStore (Map.empty))))
-        let simplified = Expr.simplify x
-        y <- checkEquiv x simplified
-        assertBoolM "Must be equal, given keccak distance axiom" y
-    ]
-  {- NOTE: These tests were designed to test behaviour on reading from a buffer such that the indices overflow 2^256.
-           However, such scenarios are impossible in the real world (the operation would run out of gas). The problem
-           is that the behaviour of bytecode interpreters does not match the semantics of SMT. Intrepreters just
-           return all zeroes for any read beyond buffer size, while in SMT reading multiple bytes may lead to overflow
-           on indices and subsequently to reading from the beginning of the buffer (wrap-around semantics).
-  , testGroup "concrete-buffer-simplification-large-index" [
-      test "copy-slice-large-index-nooverflow" $ do
-        let
-          e = CopySlice (Lit 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff) (Lit 0x0) (Lit 0x1) (ConcreteBuf "a") (ConcreteBuf "")
-          s = Expr.simplify e
-        equal <- checkEquiv e s
-        assertEqualM "Must be equal" True equal
-    , test "copy-slice-overflow-back-into-source" $ do
-        let
-          e = CopySlice (Lit 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff) (Lit 0x0) (Lit 0x2) (ConcreteBuf "a") (ConcreteBuf "")
-          s = Expr.simplify e
-        equal <- checkEquiv e s
-        assertEqualM "Must be equal" True equal
-    , test "copy-slice-overflow-beyond-source" $ do
-        let
-          e = CopySlice (Lit 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff) (Lit 0x0) (Lit 0x3) (ConcreteBuf "a") (ConcreteBuf "")
-          s = Expr.simplify e
-        equal <- checkEquiv e s
-        assertEqualM "Must be equal" True equal
-    , test "copy-slice-overflow-beyond-source-into-nonempty" $ do
-        let
-          e = CopySlice (Lit 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff) (Lit 0x0) (Lit 0x3) (ConcreteBuf "a") (ConcreteBuf "b")
-          s = Expr.simplify e
-        equal <- checkEquiv e s
-        assertEqualM "Must be equal" True equal
-    , test "read-word-overflow-back-into-source" $ do
-        let
-          e = ReadWord (Lit 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff) (ConcreteBuf "kkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkk")
-          s = Expr.simplify e
-        equal <- checkEquiv e s
-        assertEqualM "Must be equal" True equal
-  ]
-  -}
   , testGroup "ABI"
     [ testProperty "Put/get inverse" $ \x ->
         case runGetOrFail (getAbi (abiValueType x)) (runPut (putAbi x)) of
@@ -5264,49 +5054,6 @@ tests = testGroup "hevm"
   where
     (===>) = assertSolidityComputation
 
-checkEquivProp :: App m => Prop -> Prop -> m Bool
-checkEquivProp a b = fmap (fromMaybe True) $ checkEquivBase (\l r -> PNeg (PImpl l r .&& PImpl r l)) a b True
-
-checkEquivPropAndLHS :: App m => Prop -> Prop -> m Bool
-checkEquivPropAndLHS orig simp = do
-  let lhsConst = Expr.checkLHSConstProp simp
-  equiv <- checkEquivProp orig simp
-  pure $ lhsConst && equiv
-
-checkEquiv :: (Typeable a, App m) => Expr a -> Expr a -> m Bool
-checkEquiv a b = do
-  opts <- readConfig
-  if a == b then pure True else do
-    when (opts.debug) $ liftIO $ putStrLn $ "Checking equivalence of " <> show a <> " and " <> show b
-    x <- checkEquivBase (./=) a b True
-    when (opts.debug) $ liftIO $ putStrLn $ "UNSAT check, expect True: " <> show x
-    y <- checkEquivBase (.==) a b False
-    when (opts.debug) $ liftIO $ putStrLn $ "SAT check, expect False: " <> show y
-    pure $ (fromMaybe True x) && not (fromMaybe False y)
-
-checkEquivAndLHS :: (Typeable a, App m) => Expr a -> Expr a -> m Bool
-checkEquivAndLHS orig simp = do
-  opts <- readConfig
-  let lhsConst =  Expr.checkLHSConst simp
-  when (opts.debug) $ liftIO $ putStrLn $ "LHS const: " <> show lhsConst
-  equiv <- checkEquiv orig simp
-  pure $ lhsConst && equiv
-
-checkEquivBase :: (Eq a, App m) => (a -> a -> Prop) -> a -> a -> Bool -> m (Maybe Bool)
-checkEquivBase mkprop l r expect = do
-  config <- readConfig
-  let noSimplifyEnv = Env {config = config {simp = False}}
-  liftIO $ runEnv noSimplifyEnv $ do
-    withSolvers Bitwuzla 1 (Just 1) defMemLimit $ \solvers -> do
-      res <- checkSatWithProps solvers [mkprop l r]
-      let ret = case res of
-            Qed -> Just True
-            Cex {} -> Just False
-            Error _ -> Just (not expect)
-            Unknown _ -> Nothing
-      when (ret == Just (not expect)) $ liftIO $ print res
-      pure ret
-
 -- | Takes a runtime code and calls it with the provided calldata
 
 -- | Takes a creation code and some calldata, runs the creation code, and calls the resulting contract with the provided calldata
@@ -5450,477 +5197,8 @@ instance Arbitrary RLPData where
          return $ RLPData $ List [r | RLPData r <- ls])
    ]
 
-instance Arbitrary Word128 where
-  arbitrary = liftM2 fromHiAndLo arbitrary arbitrary
-
-instance Arbitrary Word160 where
-  arbitrary = liftM2 fromHiAndLo arbitrary arbitrary
-
-instance Arbitrary Word256 where
-  arbitrary = liftM2 fromHiAndLo arbitrary arbitrary
-
-instance Arbitrary W64 where
-  arbitrary = fmap W64 arbitrary
-
-instance Arbitrary W256 where
-  arbitrary = fmap W256 arbitrary
-
-instance Arbitrary Addr where
-  arbitrary = fmap Addr arbitrary
-
-instance Arbitrary (Expr EAddr) where
-  arbitrary = oneof
-    [ fmap LitAddr arbitrary
-    , fmap SymAddr (genName "addr")
-    ]
-
-instance Arbitrary (Expr Storage) where
-  arbitrary = sized genStorage
-
-instance Arbitrary (Expr EWord) where
-  arbitrary = sized defaultWord
-
-instance Arbitrary (Expr Byte) where
-  arbitrary = sized genByte
-
-newtype SymbolicJoinBytes = SymbolicJoinBytes [Expr Byte]
-  deriving (Eq, Show)
-
-instance Arbitrary SymbolicJoinBytes where
-  arbitrary = liftM SymbolicJoinBytes $ replicateM 32 arbitrary
-
-joinBytesFromList :: [Expr Byte] -> Expr EWord
-joinBytesFromList [a0, a1, a2, a3, a4, a5, a6, a7,
-                   a8, a9, a10, a11, a12, a13, a14, a15,
-                   a16, a17, a18, a19, a20, a21, a22, a23,
-                   a24, a25, a26, a27, a28, a29, a30, a31] =
-  JoinBytes a0 a1 a2 a3 a4 a5 a6 a7
-            a8 a9 a10 a11 a12 a13 a14 a15
-            a16 a17 a18 a19 a20 a21 a22 a23
-            a24 a25 a26 a27 a28 a29 a30 a31
-joinBytesFromList _ = internalError "List must contain exactly 32 elements"
-
-instance Arbitrary (Expr Buf) where
-  arbitrary = sized defaultBuf
-
-instance Arbitrary (Expr End) where
-  arbitrary = sized genEnd
-
-instance Arbitrary (ContractCode) where
-  arbitrary = oneof
-    [ fmap UnknownCode arbitrary
-    , liftM2 InitCode arbitrary arbitrary
-    , fmap RuntimeCode arbitrary
-    ]
-
-instance Arbitrary (RuntimeCode) where
-  arbitrary = oneof
-    [ fmap ConcreteRuntimeCode arbitrary
-    , fmap SymbolicRuntimeCode arbitrary
-    ]
-
-instance Arbitrary (V.Vector (Expr Byte)) where
-  arbitrary = fmap V.fromList (listOf1 arbitrary)
-
-instance Arbitrary (Expr EContract) where
-  arbitrary = sized genEContract
-
--- LitOnly
-newtype LitOnly a = LitOnly a
-  deriving (Show, Eq)
-
-newtype LitWord (sz :: Nat) = LitWord (Expr EWord)
-  deriving (Show)
-
-instance (KnownNat sz) => Arbitrary (LitWord sz) where
-  arbitrary = LitWord <$> genLit (fromInteger v)
-    where
-      v = natVal (Proxy @sz)
-
-instance Arbitrary (LitOnly (Expr Byte)) where
-  arbitrary = LitOnly . LitByte <$> arbitrary
-
-instance Arbitrary (LitOnly (Expr EWord)) where
-  arbitrary = LitOnly . Lit <$> arbitrary
-
-instance Arbitrary (LitOnly (Expr Buf)) where
-  arbitrary = LitOnly . ConcreteBuf <$> arbitrary
-
-genEContract :: Int -> Gen (Expr EContract)
-genEContract sz = do
-  c <- arbitrary
-  b <- defaultWord sz
-  n <- arbitrary
-  s <- genStorage sz
-  ts <- genStorage sz
-  pure $ C {code=c, storage=s, tStorage=ts, balance=b, nonce=n}
-
--- ZeroDepthWord
-newtype ZeroDepthWord = ZeroDepthWord (Expr EWord)
-  deriving (Show, Eq)
-
-instance Arbitrary ZeroDepthWord where
-  arbitrary = do
-    fmap ZeroDepthWord . sized $ genWord 0
-
--- WriteWordBuf
-newtype WriteWordBuf = WriteWordBuf (Expr Buf)
-  deriving (Show, Eq)
-
-instance Arbitrary WriteWordBuf where
-  arbitrary = do
-    let mkBuf = oneof
-          [ pure $ ConcreteBuf ""       -- empty
-          , fmap ConcreteBuf arbitrary  -- concrete
-          , sized (genBuf 100)          -- overlapping writes
-          , arbitrary                   -- sparse writes
-          ]
-    fmap WriteWordBuf mkBuf
-
--- GenCopySliceBuf
-newtype GenCopySliceBuf = GenCopySliceBuf (Expr Buf)
-  deriving (Show, Eq)
-
-instance Arbitrary GenCopySliceBuf where
-  arbitrary = do
-    let mkBuf = oneof
-          [ pure $ ConcreteBuf ""
-          , fmap ConcreteBuf arbitrary
-          , arbitrary
-          ]
-    fmap GenCopySliceBuf mkBuf
-
--- GenWriteStorageExpr
-newtype GenWriteStorageExpr = GenWriteStorageExpr (Expr EWord, Expr Storage)
-  deriving (Show, Eq)
-
-instance Arbitrary GenWriteStorageExpr where
-  arbitrary = do
-    slot <- arbitrary
-    let mkStore = oneof
-          [ pure $ ConcreteStore mempty
-          , fmap ConcreteStore arbitrary
-          , do
-              -- generate some write chains where we know that at least one
-              -- write matches either the input addr, or both the input
-              -- addr and slot
-              let addWrites :: Expr Storage -> Int -> Gen (Expr Storage)
-                  addWrites b 0 = pure b
-                  addWrites b n = liftM3 SStore arbitrary arbitrary (addWrites b (n - 1))
-              s <- arbitrary
-              addMatch <- fmap (SStore slot) arbitrary
-              let withMatch = addMatch s
-              newWrites <- oneof [ pure 0, pure 1, fmap (`mod` 5) arbitrary ]
-              addWrites withMatch newWrites
-          , arbitrary
-          ]
-    store <- mkStore
-    pure $ GenWriteStorageExpr (slot, store)
-
--- GenWriteByteIdx
-newtype GenWriteByteIdx = GenWriteByteIdx (Expr EWord)
-  deriving (Show, Eq)
-
-instance Arbitrary GenWriteByteIdx where
-  arbitrary = do
-    -- 1st: can never overflow an Int
-    -- 2nd: can overflow an Int
-    let mkIdx = frequency [ (10, genLit 1_000_000) , (1, fmap Lit arbitrary) ]
-    fmap GenWriteByteIdx mkIdx
-
-newtype LitProp = LitProp Prop
-  deriving (Show, Eq)
-
-instance Arbitrary LitProp where
-  arbitrary = LitProp <$> sized (genProp True)
-
-
-newtype StorageExp = StorageExp (Expr EWord)
-  deriving (Show, Eq)
-
-instance Arbitrary StorageExp where
-  arbitrary = StorageExp <$> (genStorageExp)
-
-genStorageExp :: Gen (Expr EWord)
-genStorageExp = do
-  fromPos <- genSlot
-  storage <- genStorageWrites
-  pure $ SLoad fromPos storage
-
-genSlot :: Gen (Expr EWord)
-genSlot = frequency [ (1, do
-                        buf <- genConcreteBufSlot 64
-                        case buf of
-                          (ConcreteBuf b) -> do
-                            key <- genLit 10
-                            pure $ Expr.MappingSlot b key
-                          _ -> internalError "impossible"
-                        )
-                     -- map element
-                     ,(2, do
-                        l <- genLit 10
-                        buf <- genConcreteBufSlot 64
-                        pure $ Add (Keccak buf) l)
-                    -- Array element
-                     ,(2, do
-                        l <- genLit 10
-                        buf <- genConcreteBufSlot 32
-                        pure $ Add (Keccak buf) l)
-                     -- member of the Contract
-                     ,(2, pure $ Lit 20)
-                     -- array element
-                     ,(2, do
-                        arrayNum :: Int <- arbitrary
-                        offs :: W256 <- arbitrary
-                        pure $ Lit $ fst (Expr.preImages !! (arrayNum `mod` 3)) + (offs `mod` 3))
-                     -- random stuff
-                     ,(1, pure $ Lit (maxBound :: W256))
-                     ]
-
--- Generates an N-long buffer, all with the same value, at most 8 different ones
-genConcreteBufSlot :: Int -> Gen (Expr Buf)
-genConcreteBufSlot len = do
-  b :: Word8 <- arbitrary
-  pure $ ConcreteBuf $ BS.pack ([ 0 | _ <- [0..(len-2)]] ++ [b])
-
-genStorageWrites :: Gen (Expr Storage)
-genStorageWrites = do
-  toSlot <- genSlot
-  val <- genLit (maxBound :: W256)
-  store <- frequency [ (3, pure $ AbstractStore (SymAddr "") Nothing)
-                     , (2, genStorageWrites)
-                     ]
-  pure $ SStore toSlot val store
-
-instance Arbitrary Prop where
-  arbitrary = sized (genProp False)
-
-genProps :: Bool -> Int -> Gen [Prop]
-genProps onlyLits sz2 = listOf $ genProp onlyLits sz2
-
-genProp :: Bool -> Int -> Gen (Prop)
-genProp _ 0 = PBool <$> arbitrary
-genProp onlyLits sz = oneof
-  [ liftM2 PEq subWord subWord
-  , liftM2 PLT subWord subWord
-  , liftM2 PGT subWord subWord
-  , liftM2 PLEq subWord subWord
-  , liftM2 PGEq subWord subWord
-  , fmap PNeg subProp
-  , liftM2 PAnd subProp subProp
-  , liftM2 POr subProp subProp
-  , liftM2 PImpl subProp subProp
-  ]
-  where
-    subWord = if onlyLits then frequency [(2, Lit <$> arbitrary)
-                                         ,(1, pure $ Lit 0)
-                                         ,(1, pure $ Lit Expr.maxLit)
-                                         ]
-                          else genWord 1 (sz `div` 2)
-    subProp = genProp onlyLits (sz `div` 2)
-
-genByte :: Int -> Gen (Expr Byte)
-genByte 0 = fmap LitByte arbitrary
-genByte sz = oneof
-  [ liftM2 IndexWord subWord subWord
-  , liftM2 ReadByte subWord subBuf
-  ]
-  where
-    subWord = defaultWord (sz `div` 10)
-    subBuf = defaultBuf (sz `div` 10)
-
-genLit :: W256 -> Gen (Expr EWord)
-genLit bound = do
-  w <- arbitrary
-  pure $ Lit (w `mod` bound)
-
 genNat :: Gen Int
 genNat = fmap unsafeInto (arbitrary :: Gen Natural)
-
-genName :: String -> Gen Text
--- In order not to generate SMT reserved words, we prepend with "esc_"
-genName ty = fmap (T.pack . (("esc_" <> ty <> "_") <> )) $ listOf1 (oneof . (fmap pure) $ ['a'..'z'] <> ['A'..'Z'])
-
-genEnd :: Int -> Gen (Expr End)
-genEnd 0 = oneof
-  [ fmap (Failure mempty mempty . UnrecognizedOpcode) arbitrary
-  , pure $ Failure mempty mempty IllegalOverflow
-  , pure $ Failure mempty mempty SelfDestruction
-  ]
-genEnd sz = oneof
-  [ liftM3 Failure subProp (pure mempty) (fmap Revert subBuf)
-  , liftM4 Success subProp (pure mempty) subBuf arbitrary
-  -- TODO Partial
-  ]
-  where
-    subBuf = defaultBuf (sz `div` 2)
-    subProp = genProps False (sz `div` 2)
-
-genWord :: Int -> Int -> Gen (Expr EWord)
-genWord litFreq 0 = frequency
-  [ (litFreq, do
-      val <- frequency
-       [ (10, fmap (`mod` 100) arbitrary)
-       , (1, pure 0)
-       , (1, pure Expr.maxLit)
-       , (1, arbitrary)
-       ]
-      pure $ Lit val
-    )
-  , (1, oneof
-      [ pure Origin
-      , pure Coinbase
-      , pure Timestamp
-      , pure BlockNumber
-      , pure PrevRandao
-      , pure GasLimit
-      , pure ChainId
-      , pure BaseFee
-      --, liftM2 SelfBalance arbitrary arbitrary
-      --, liftM2 Gas arbitrary arbitrary
-      , fmap Lit arbitrary
-      , fmap joinBytesFromList $ replicateM 32 arbitrary
-      , fmap Var (genName "word")
-      ]
-    )
-  ]
-genWord litFreq sz = frequency
-  [ (litFreq, do
-      val <- frequency
-       [ (10, fmap (`mod` 100) arbitrary)
-       , (1, arbitrary)
-       ]
-      pure $ Lit val
-    )
-  , (1, oneof
-    [ liftM2 Add subWord subWord
-    , liftM2 Sub subWord subWord
-    , liftM2 Mul subWord subWord
-    , liftM2 Div subWord subWord
-    , liftM2 SDiv subWord subWord
-    , liftM2 Mod subWord subWord
-    , liftM2 SMod subWord subWord
-    --, liftM3 AddMod subWord subWord subWord
-    --, liftM3 MulMod subWord subWord subWord -- it works, but it's VERY SLOW
-    --, liftM2 Exp subWord litWord
-    , liftM2 SEx subWord subWord
-    , liftM2 Min subWord subWord
-    , liftM2 LT subWord subWord
-    , liftM2 GT subWord subWord
-    , liftM2 LEq subWord subWord
-    , liftM2 GEq subWord subWord
-    , liftM2 SLT subWord subWord
-    --, liftM2 SGT subWord subWord
-    , liftM2 Eq subWord subWord
-    , fmap IsZero subWord
-    , liftM2 And subWord subWord
-    , liftM2 Or subWord subWord
-    , liftM2 Xor subWord subWord
-    , fmap Not subWord
-    , liftM2 SHL subWord subWord
-    , liftM2 SHR subWord subWord
-    , liftM2 SAR subWord subWord
-    , fmap BlockHash subWord
-    --, liftM3 Balance arbitrary arbitrary subWord
-    --, fmap CodeSize subWord
-    --, fmap ExtCodeHash subWord
-    , fmap Keccak subBuf
-    , fmap SHA256 subBuf
-    , liftM2 SLoad subWord subStore
-    , liftM2 ReadWord genReadIndex subBuf
-    , fmap BufLength subBuf
-    , do
-      one <- subByte
-      two <- subByte
-      three <- subByte
-      four <- subByte
-      five <- subByte
-      six <- subByte
-      seven <- subByte
-      eight <- subByte
-      nine <- subByte
-      ten <- subByte
-      eleven <- subByte
-      twelve <- subByte
-      thirteen <- subByte
-      fourteen <- subByte
-      fifteen <- subByte
-      sixteen <- subByte
-      seventeen <- subByte
-      eighteen <- subByte
-      nineteen <- subByte
-      twenty <- subByte
-      twentyone <- subByte
-      twentytwo <- subByte
-      twentythree <- subByte
-      twentyfour <- subByte
-      twentyfive <- subByte
-      twentysix <- subByte
-      twentyseven <- subByte
-      twentyeight <- subByte
-      twentynine <- subByte
-      thirty <- subByte
-      thirtyone <- subByte
-      thirtytwo <- subByte
-      pure $ JoinBytes
-        one two three four five six seven eight nine ten
-        eleven twelve thirteen fourteen fifteen sixteen
-        seventeen eighteen nineteen twenty twentyone
-        twentytwo twentythree twentyfour twentyfive
-        twentysix twentyseven twentyeight twentynine
-        thirty thirtyone thirtytwo
-    ])
-  ]
- where
-   subWord = genWord litFreq (sz `div` 5)
-   subBuf = defaultBuf (sz `div` 10)
-   subStore = genStorage (sz `div` 10)
-   subByte = genByte (sz `div` 10)
-   genReadIndex = do
-    o :: (Expr EWord) <- subWord
-    pure $ case o of
-      Lit w -> Lit $ w `mod` into (maxBound :: Word64)
-      _ -> o
-
-genWordArith :: Int -> Int -> Gen (Expr EWord)
-genWordArith litFreq 0 = frequency
-  [ (litFreq, fmap Lit arbitrary)
-  , (1, oneof [ fmap Lit arbitrary ])
-  ]
-genWordArith litFreq sz = frequency
-  [ (litFreq, fmap Lit arbitrary)
-  , (20, frequency
-    [ (20, liftM2 Add  subWord subWord)
-    , (20, liftM2 Sub  subWord subWord)
-    , (20, liftM2 Mul  subWord subWord)
-    , (20, liftM2 SEx  subWord subWord)
-    , (20, liftM2 Xor  subWord subWord)
-    -- these reduce variability
-    , (3 , liftM2 Min  subWord subWord)
-    , (3 , liftM2 Div  subWord subWord)
-    , (3 , liftM2 SDiv subWord subWord)
-    , (3 , liftM2 Mod  subWord subWord)
-    , (3 , liftM2 SMod subWord subWord)
-    , (3 , liftM2 SHL  subWord subWord)
-    , (3 , liftM2 SHR  subWord subWord)
-    , (3 , liftM2 SAR  subWord subWord)
-    , (3 , liftM2 Or   subWord subWord)
-    -- comparisons, reducing variability greatly
-    , (1 , liftM2 LEq  subWord subWord)
-    , (1 , liftM2 GEq  subWord subWord)
-    , (1 , liftM2 SLT  subWord subWord)
-    --(1, , liftM2 SGT subWord subWord
-    , (1 , liftM2 Eq   subWord subWord)
-    , (1 , liftM2 And  subWord subWord)
-    , (1 , fmap IsZero subWord        )
-    -- Expensive below
-    --(1,  liftM3 AddMod subWord subWord subWord
-    --(1,  liftM3 MulMod subWord subWord subWord
-    --(1,  liftM2 Exp subWord litWord
-    ])
-  ]
- where
-   subWord = genWordArith (litFreq `div` 2) (sz `div` 2)
 
 
 -- Finds SLoad -> SStore. This should not occur in most scenarios
@@ -5930,93 +5208,6 @@ badStoresInExpr exprs = any (getAny . foldExpr match mempty) exprs
   where
       match (SLoad _ (SStore _ _ _)) = Any True
       match _ = Any False
-
-defaultBuf :: Int -> Gen (Expr Buf)
-defaultBuf = genBuf (4_000_000)
-
-defaultWord :: Int -> Gen (Expr EWord)
-defaultWord = genWord 10
-
-maybeBoundedLit :: W256 -> Gen (Expr EWord)
-maybeBoundedLit bound = do
-  o <- (arbitrary :: Gen (Expr EWord))
-  pure $ case o of
-        Lit w -> Lit $ w `mod` bound
-        _ -> o
-
-genBuf :: W256 -> Int -> Gen (Expr Buf)
-genBuf _ 0 = oneof
-  [ fmap AbstractBuf (genName "buf")
-  , fmap ConcreteBuf arbitrary
-  ]
-genBuf bound sz = oneof
-  [ liftM3 WriteWord (maybeBoundedLit bound) subWord subBuf
-  , liftM3 WriteByte (maybeBoundedLit bound) subByte subBuf
-  -- we don't generate copyslice instances where:
-  --   - size is abstract
-  --   - size > 100 (due to unrolling in SMT.hs)
-  --   - literal dstOffsets are > 4,000,000 (due to unrolling in SMT.hs)
-  -- n.b. that 4,000,000 is the theoretical maximum memory size given a 30,000,000 block gas limit
-  , liftM5 CopySlice genReadIndex (maybeBoundedLit bound) smolLitWord subBuf subBuf
-  ]
-  where
-    -- copySlice gets unrolled in the generated SMT so we can't go too crazy here
-    smolLitWord = do
-      w <- arbitrary
-      pure $ Lit (w `mod` 100)
-    subWord = defaultWord (sz `div` 5)
-    subByte = genByte (sz `div` 10)
-    subBuf = genBuf bound (sz `div` 10)
-    genReadIndex = do
-      o :: (Expr EWord) <- subWord
-      pure $ case o of
-        Lit w -> Lit $ w `mod` into (maxBound :: Word64)
-        _ -> o
-
-genStorage :: Int -> Gen (Expr Storage)
-genStorage 0 = oneof
-  [ liftM2 AbstractStore arbitrary (pure Nothing)
-  , fmap ConcreteStore $ resize 5 arbitrary
-  ]
-genStorage sz = liftM3 SStore key val subStore
-  where
-    subStore = genStorage (sz `div` 10)
-    val = defaultWord (sz `div` 5)
-    key = genStorageKey
-
-genStorageKey :: Gen (Expr EWord)
-genStorageKey = frequency
-     -- array slot
-    [ (4, liftM2 Expr.ArraySlotWithOffs (genByteStringKey 32) (genLit 5))
-    , (4, fmap Expr.ArraySlotZero (genByteStringKey 32))
-     -- mapping slot
-    , (8, liftM2 Expr.MappingSlot (genByteStringKey 64) (genLit 5))
-     -- small slot
-    , (4, genLit 20)
-    -- unrecognized slot type
-    , (1, genLit 5)
-    ]
-
-genByteStringKey :: W256 -> Gen (ByteString)
-genByteStringKey len = do
-  b :: Word8 <- arbitrary
-  pure $ BS.pack ([ 0 | _ <- [0..(len-2)]] ++ [b `mod` 5])
-
--- GenWriteStorageLoad
-newtype GenWriteStorageLoad = GenWriteStorageLoad (Expr EWord)
-  deriving (Show, Eq)
-
-instance Arbitrary GenWriteStorageLoad where
-  arbitrary = do
-    load <- genStorageLoad 10
-    pure $ GenWriteStorageLoad load
-
-    where
-      genStorageLoad :: Int -> Gen (Expr EWord)
-      genStorageLoad sz = liftM2 SLoad key subStore
-        where
-          subStore = genStorage (sz `div` 10)
-          key = genStorageKey
 
 data Invocation
   = SolidityCall Text [AbiValue]
